Modified to allow a broader set of analysis options:
Streamlined debug/development mode vs normal operation selection
Can filter on a dataset column field as passed in from the command line.
Allow the category to group the results by to be passed from teh command line.
Generalized the reporting function to allow it to handle any category as long
as a list of entries in the category and a category name lookup object are
provided.
Streamlined the selection of the maximum number of items to report on.